move validator server impl
