+++ crates/sui-network/Cargo.toml
@@ -18,6 +18,13 @@ tonic.workspace = true
 dashmap.workspace = true
 tower.workspace = true
 shared-crypto.workspace = true
+eyre.workspace = true
+http.workspace = true
+tokio-rustls.workspace = true
+tonic-health.workspace = true
+tower-http.workspace = true
+sui-http.workspace = true
+sui-tls.workspace = true
 
 sui-macros.workspace = true
 sui-storage.workspace = true
