+++ crates/sui-network/src/lib.rs
@@ -9,6 +9,7 @@ pub mod discovery;
 pub mod randomness;
 pub mod state_sync;
 pub mod utils;
+pub mod validator;
 
 pub use tonic;
