+++ crates/sui-network/src/validator/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod server;
+
+pub use server::{Server, ServerBuilder, SUI_TLS_SERVER_NAME};
