+++ crates/sui-node/Cargo.toml
@@ -52,8 +52,8 @@ sui-types.workspace = true
 mysten-metrics.workspace = true
 mysten-service.workspace = true
 mysten-common.workspace = true
-typed-store.workspace = true
 mysten-network.workspace = true
+typed-store.workspace = true
 telemetry-subscribers.workspace = true
 fastcrypto.workspace = true
 fastcrypto-zkp.workspace = true
